requirements.txt
Flask==2.3.3
httpx==0.24.1
python-dotenv==1.1.0
